refactor(portaltransform): 重构物品转换配方相关代码
-简化 Dimensions 类中的 current 和 target 方法
- 重命名 ItemTransform 相关的包结构- 优化物品转换配方的查找和处理逻辑
- 移除未使用的代码和注释
